lpp/requirements.txt
shap
xgboost
numpy
scikit-learn
sklearn-pandas
tqdm
tldextract
requests
ratelimit
pymongo
pandas
aiohttp
asyncio
fastapi==0.62.0
uvicorn==0.13.1
joblib
celery[redis]
threadpoolctl
librabbitmq
amqp
urllib3
gevent
kombu
joblib
importlib-metadata
billiard
starlette
urllib3
vine
